requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
opencv-python>=4.5.0
Pillow>=9.0.0
tqdm>=4.60.0
ipywidgets>=7.0.0
scikit-image>=0.19.0
imageio>=2.15.0
import-ipynb>=0.2.0